更改项目结构， 三个页面内容仔细重绘，并加入了按钮事件等功能， 页面加入了参数调试显示功能 完成了大致雏形
