--- src/pages/index/index.skeleton.wxml
@@ -1,43 +1,0 @@
-<!--
-此文件为开发者工具生成，生成时间: 2020/9/19 下午7:18:11 
-使用方法：
-在 pages\index\index.wxml 引入模板
-
-```
-<import src="index.skeleton.wxml"/>
-<template is="skeleton" wx-if="{{loading}}" />
-```
-
-在 pages\index\index.wxss 中引入样式
-```
-@import "./index.skeleton.wxss";
-```
-
-更多详细信息可以参考文档：https://developers.weixin.qq.com/miniprogram/dev/devtools/skeleton.html
--->
-<template name="skeleton">
-  <view class="sk-container">
-    <view class="container">
-      <view class="userinfo">
-        <image class="userinfo-avatar sk-image" mode="cover">
-        </image>
-        <text class="userinfo-nickname sk-transparent sk-text-14-2857-519 sk-text">
-          花样作死小王子
-        </text>
-      </view>
-      <view class="usermotto">
-        <text class="user-motto sk-transparent sk-text-14-2857-69 sk-text">
-          Hello World
-        </text>
-      </view>
-      <view class="btn-area">
-        <navigator class="sk-transparent sk-text-14-2857-400 sk-text">
-          跳转到ui2页面
-        </navigator>
-        <view class="mt20 list-item sk-transparent sk-text-28-9474-304 sk-text" data-page="ui2">
-          ScrollLine
-        </view>
-      </view>
-    </view>
-  </view>
-</template>
