--- src/pages/index/index.skeleton.wxss
@@ -1,52 +1,0 @@
-/*
-此文件为开发者工具生成，生成时间: 2020/9/19 下午7:18:11 
-
-在 pages\index\index.wxss 中引入样式
-```
-@import "./index.skeleton.wxss";
-```
-
-更多详细信息可以参考文档：https://developers.weixin.qq.com/miniprogram/dev/devtools/skeleton.html
-*/
-.sk-transparent {
-    color: transparent !important;
-  }
-.sk-text-14-2857-519 {
-    background-image: linear-gradient(transparent 14.2857%, #EEEEEE 0%, #EEEEEE 85.7143%, transparent 0%) !important;
-    background-size: 100% 35.0000rpx;
-    position: relative !important;
-  }
-.sk-text {
-    background-origin: content-box !important;
-    background-clip: content-box !important;
-    background-color: transparent !important;
-    color: transparent !important;
-    background-repeat: repeat-y !important;
-  }
-.sk-text-14-2857-69 {
-    background-image: linear-gradient(transparent 14.2857%, #EEEEEE 0%, #EEEEEE 85.7143%, transparent 0%) !important;
-    background-size: 100% 35.0000rpx;
-    position: relative !important;
-  }
-.sk-text-14-2857-400 {
-    background-image: linear-gradient(transparent 14.2857%, #EEEEEE 0%, #EEEEEE 85.7143%, transparent 0%) !important;
-    background-size: 100% 35.0000rpx;
-    position: relative !important;
-  }
-.sk-text-28-9474-304 {
-    background-image: linear-gradient(transparent 28.9474%, #EEEEEE 0%, #EEEEEE 71.0526%, transparent 0%) !important;
-    background-size: 100% 59.3750rpx;
-    position: relative !important;
-  }
-.sk-image {
-    background: #EFEFEF !important;
-  }
-.sk-container {
-    position: absolute;
-    left: 0;
-    top: 0;
-    width: 100%;
-    height: 100%;
-    overflow: hidden;
-    background-color: transparent;
-  }
