--- src/pages/index/index.wxss
@@ -1,31 +1,0 @@
-/**index.wxss**/
-.userinfo {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
-
-.userinfo-avatar {
-  width: 128rpx;
-  height: 128rpx;
-  margin: 20rpx;
-  border-radius: 50%;
-}
-
-.userinfo-nickname {
-  color: #aaa;
-}
-
-.usermotto {
-  margin-top: 200px;
-}
-
-.mt20 {
-  margin-top: 20rpx;
-}
-
-.list-item {
-  height: 60rpx;
-  line-height: 60rpx;
-  text-indent: 30rpx;
-}
