--- src/pages/input/input.json
@@ -1,3 +1,0 @@
-{
-  "usingComponents": {}
-}
