--- src/pages/input/input.wxss
@@ -1,1 +1,0 @@
-/* pages/input/input.wxss */
