+++ src/pages/p0/p0.wxml
@@ -2,6 +2,9 @@
 <view>调试数据</view>
 <view>{{currencyIndexB}}</view>
 <view>{{currencyIndexE}}</view>
+<view>{{inputValB}}</view>
+<view>{{inputValE}}</view>
+<view>{{showTopTips}}</view>
 
 <view class="page" data-weui-theme="{{theme}}">
     <view class="page__hd">
@@ -15,54 +18,47 @@
         <!-- 币种选择 -->
         <view class="weui-cells__title">币种选择</view>
         <view class="weui-cells weui-cells_after-title">
-                <!-- <view class="weui-cell weui-ell_active weui-cell_select">
-                    <view class="weui-cell__bd">
-                        <picker bindchange="bindAccountChange" value="{{accountIndex}}" range="{{accounts}}">
-                            <view class="weui-select">{{accounts[accountIndex]}}</view>
-                        </picker>
-                    </view>
-                </view>c -->
-                <view class="weui-cell weui-cell_active">
-                    <view class="weui-cell__hd">
-                        <view class="weui-label">置换前</view>
-                    </view>
-                    <view class="weui-cell__bd">
-                        <picker bindchange="bindCurrencyChangeB" value="{{currencyIndexB}}" range="{{currenciesZH}}">
-                            <view class="weui-select weui-select_in-select-after">{{currenciesZH[currencyIndexB]}}</view>
-                        </picker>
-                    </view>
-                    <view class="weui-cell__hd">
-                        <view class="weui-label">置换后</view>
-                    </view>
-                    <view class="weui-cell__bd">
-                        <picker bindchange="bindCurrencyChangeE" value="{{currencyIndexE}}" range="{{currenciesZH}}">
-                            <view class="weui-select weui-select_in-select-after">{{currenciesZH[currencyIndexE]}}</view>
-                        </picker>
-                    </view>
+            <view class="weui-cell weui-cell_active">
+                <view class="weui-cell__hd">
+                    <view class="weui-label">置换前</view>
                 </view>
+                <view class="weui-cell__bd">
+                    <picker bindchange="bindCurrencyChangeB" value="{{currencyIndexB}}" range="{{currenciesZH}}">
+                        <view class="weui-select weui-select_in-select-after">{{currenciesZH[currencyIndexB]}}</view>
+                    </picker>
+                </view>
+                <view class="weui-cell__hd">
+                    <view class="weui-label">置换后</view>
+                </view>
+                <view class="weui-cell__bd">
+                    <picker bindchange="bindCurrencyChangeE" value="{{currencyIndexE}}" range="{{currenciesZH}}">
+                        <view class="weui-select weui-select_in-select-after">{{currenciesZH[currencyIndexE]}}</view>
+                    </picker>
+                </view>
+            </view>
         </view>
 
         <view class="weui-btn-area">
             <!-- <button class="weui-btn" type="primary" bindtap="showTopTips">确定</button> -->
-            <button class="weui-btn" type="primary">查询</button>
+            <button class="weui-btn" type="primary" bindtap="showTopTips">查询</button>
         </view>
 
-        <view class="weui-cells__title">文本框</view>
+        <view class="weui-cells__title">兑换换算</view>
         <view class="weui-cells weui-cells_after-title">
             <view class="weui-cell weui-cell_active">
                 <view class="weui-cell__bd">
-                    <input class="weui-input" placeholder="请输入金额" />
+                    <input class="weui-input" bindchange="bindInputChangeB" type="number" value="{{inputValB}}" placeholder="请输入金额" />
+					<!-- <input class="weui-input" type="number" placeholder="请输入金额" value="12345678907" placeholder-class="weui-input__placeholder" /> -->
                 </view>
             </view>
         </view>
-        <view class="weui-cells__title">文本框</view>
+        <!-- <view class="weui-cells__title">文本框</view> -->
         <view class="weui-cells weui-cells_after-title">
             <view class="weui-cell weui-cell_active">
                 <view class="weui-cell__bd">
-                    <input class="weui-input" placeholder="兑换的金额" />
+                    <input class="weui-input" bindinput="bindInputChangeE" type="number" value="{{inputValB}}" placeholder="兑换的金额" />
                 </view>
             </view>
         </view>
-
     </view>
 </view>
