+++ src/pages/p2/p2.wxml
@@ -1,2 +1,19 @@
 <!--pages/p2/p2.wxml-->
-<text>pages/p2/p2.wxml</text>
+
+<view class="page" data-weui-theme="{{theme}}">
+    <view class="page__hd">
+        <view class="page__title">垃圾</view>
+        <view class="page__desc">垃圾</view>
+        <view class="page__desc">
+            <navigator url="/example/form/form" class="link">垃圾</navigator> 。
+        </view>
+    </view>
+    <view class="page__bd">
+        <view class="button-sp-area cell">
+            <a class="weui-btn_cell weui-btn_cell-default" bindtap="gotoPage" data-url="/pages/webview/webview?srcurl=https://github.com/Louis-J/HUILV">程序说明</a>
+            <a class="weui-btn_cell weui-btn_cell-default" bindtap="gotoPage" data-url="/pages/logs/logs" >LOG信息</a>
+            <a class="weui-btn_cell weui-btn_cell-primary">关于我们</a>
+        </view>
+
+    </view>
+</view>
