+++ src/pages/webview/webview.wxml
@@ -1,0 +1,1 @@
+<web-view src="{{srcurl}}"></web-view>
